feat: Create SalesService Authorization + company logic (#44)

+++ src/Services/SalesService/SalesService.Application/Commands/Products/DeleteProduct.cs
@@ -4,7 +4,10 @@ using Shared.Exceptions;
 
 namespace SalesService.Application.Commands.Products
 {
-    public record DeleteProductCommand(string ProductId) : IRequest<Unit>;
+    public record DeleteProductCommand(string ProductId) : IRequest<Unit>, IHaveCompanyId
+    {
+        public string CompanyId { get; set; } = string.Empty;
+    }
 
     public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Unit>
     {
@@ -20,6 +23,9 @@ namespace SalesService.Application.Commands.Products
             var existingProduct = await _productRepository.GetProductByIdAsync(request.ProductId);
             if (existingProduct == null)
                 throw new NotFoundException($"Product with id {request.ProductId} not found");
+
+            if (existingProduct.CompanyId != request.CompanyId)
+                throw new ForbiddenException("You are not able to delete product from other company than you are register in");
 
             await _productRepository.DeleteProductAsync(existingProduct);
