+++ src/Services/SalesService/SalesService.Application/Commands/Services/DeleteService.cs
@@ -4,7 +4,10 @@ using Shared.Exceptions;
 
 namespace SalesService.Application.Commands.Services
 {
-    public record DeleteServiceCommand(string ServiceId) : IRequest<Unit>;
+    public record DeleteServiceCommand(string ServiceId) : IRequest<Unit>, IHaveCompanyId
+    {
+        public string CompanyId { get; set; } = string.Empty;
+    }
 
     public class DeleteServiceCommandHandler : IRequestHandler<DeleteServiceCommand, Unit>
     {
@@ -20,6 +23,9 @@ namespace SalesService.Application.Commands.Services
             var existingService = await _serviceRepository.GetServiceByIdAsync(request.ServiceId);
             if (existingService == null)
                 throw new NotFoundException($"Service with id {request.ServiceId} not found");
+
+            if (existingService.CompanyId != request.CompanyId)
+                throw new ForbiddenException("You are not able to delete service from other company than you are register in");
 
             await _serviceRepository.DeleteServiceAsync(existingService);
